+++ lib/src/ffmpeg/windows_ffmpeg.dart
@@ -19,6 +19,10 @@ class WindowsFfmpeg extends BaseFfmpeg {
   Future<void> createFfmpeg() async {
     final dir = await path();
     final file = File('$dir${Platform.pathSeparator}$ffmpegName');
+    if (file.existsSync()) {
+      debugPrint('$ffmpegName is already exists.');
+      return;
+    }
 
     await writeFile(file, 'assets/$ffmpegName');
